feat: Implement dashboard page with analysis and execution views
- Added DashboardPageContent component to display the main dashboard layout.
- Integrated DashboardChipView, DashboardAnalysisGrid, and DashboardExecutionView components for chip experiments, analysis overview, and execution history respectively.
- Created ExecutionPageContent component to manage execution history and details with state management and data fetching.
- Developed DashboardAnalysisWrapper, DashboardChipWrapper, and DashboardExecutionWrapper components for consistent styling and navigation.
- Introduced EmbeddedAnalysisView and CompactAnalysisView for embedding analysis views with navigation and loading states.
- Implemented responsive design and loading indicators for better user experience.
